requirements.txt
requests
pyyaml
sgp4
numpy
